Get Mock stuff compiling again, add missing unit test related nomic package to package.json

+++ protocol/contracts/libraries/Silo/LibWhitelist.sol
@@ -31,7 +31,7 @@ library LibWhitelist {
     event WhitelistToken(
         address indexed token,
         bytes4 selector,
-        uint256 stalkPerBdvPerSeason,
+        uint32 stalkPerBdvPerSeason,
         uint256 stalk
     );
